# openapi.yaml
openapi: 3.0.3
info:
  title: Go RBAC + H3 + Simulation API (bare minimum)
  version: "1.0.0"
servers:
  - url: http://localhost:8080
paths:
  /auth/register:
    post:
      summary: Register user (role defaults to user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
  /auth/login:
    post:
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: Unauthorized
  /cells:
    post:
      security:
        - bearerAuth: []
      summary: Create cell (stores computed H3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCellRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  h3: { type: string }
        "401":
          description: Unauthorized
  /cells/{h3}:
    get:
      security:
        - bearerAuth: []
      summary: Get cells by exact H3 index
      parameters:
        - in: path
          name: h3
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /cells/nearby:
    get:
      security:
        - bearerAuth: []
      summary: Query cells in k-ring around a center H3
      parameters:
        - in: query
          name: h3
          required: true
          schema: { type: string }
        - in: query
          name: k
          required: false
          schema: { type: integer, default: 1, minimum: 0, maximum: 5 }
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /simulation/run:
    post:
      security:
        - bearerAuth: []
      summary: Start event-driven simulation (enqueues TICK events)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunSimulationRequest'
      responses:
        "202":
          description: Accepted
        "401":
          description: Unauthorized
  /simulation/runs/{id}:
    get:
      security:
        - bearerAuth: []
      summary: Get simulation run status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /analytics/recompute:
    post:
      security:
        - bearerAuth: []
      summary: Recompute aggregated analytics table (cells count by H3 parent)
      parameters:
        - in: query
          name: res
          required: false
          schema: { type: integer, default: 7 }
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /analytics:
    get:
      security:
        - bearerAuth: []
      summary: List aggregated counts by H3 parent
      parameters:
        - in: query
          name: res
          required: false
          schema: { type: integer, default: 7 }
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 200, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /audit:
    get:
      security:
        - bearerAuth: []
      summary: List audit logs (admin only)
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: user@example.com }
        password: { type: string, example: pass1234 }
        role: { type: string, enum: [user, admin], example: user }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
    CreateCellRequest:
      type: object
      required: [lat, lng]
      properties:
        name: { type: string, example: my-cell }
        lat: { type: number, format: double, example: 43.238949 }
        lng: { type: number, format: double, example: 76.889709 }
        resolution: { type: integer, example: 9 }
    RunSimulationRequest:
      type: object
      properties:
        ticks: { type: integer, example: 5, minimum: 1, maximum: 100 }